requirements.txt
arrow
click
daiquiri
duckdb
easier
flake8
fleming
holoviews[recommended]
hvplot
ibis-framework[duckdb]
ibis-framework[postgres]
ipython
jupyter
jupyterlab
numpy
pandas
pdbpp
pgcli
pip
psutil
psycopg2
pytest
python-dateutil
scipy
streamlit
streamlit-folium
